+++ backend/api/views/__init__.py
@@ -1,0 +1,7 @@
+from api.views.BoardViewSet import BoardViewSet
+from api.views.ParticipantViewSet import ParticipantViewSet
+from api.views.PostViewSet import PostViewSet
+from api.views.ThreadViewSet import ThreadViewSet
+from api.views.UserViewSet import UserViewSet
+
+__all__ = ['BoardViewSet', 'ParticipantViewSet', 'PostViewSet', 'ThreadViewSet', 'UserViewSet']
